refactor: remove UpdateVaultProvider and integrate updateVault functionality directly into WriteStorageProvider for streamlined vault management across desktop and extension clients
