+++ clients/extension/src/vault/state/vaults.ts
@@ -2,7 +2,7 @@ import { vaultsQueryKey } from '@core/ui/query/keys'
 import { Vault } from '@core/ui/vault/Vault'
 
 import { getPersistentState } from '../../state/persistent/getPersistentState'
-import { usePersistentStateMutation } from '../../state/persistent/usePersistentStateMutation'
+import { setPersistentState } from '../../state/persistent/setPersistentState'
 import { usePersistentStateQuery } from '../../state/persistent/usePersistentStateQuery'
 
 const initialValue: Vault[] = []
@@ -11,8 +11,8 @@ const [key] = vaultsQueryKey
 
 export const getVaults = async () => getPersistentState(key, initialValue)
 
-export const useVaultsMutation = () => {
-  return usePersistentStateMutation<Vault[]>(key)
+export const updateVaults = async (vaults: Vault[]) => {
+  await setPersistentState(key, vaults)
 }
 
 export const useVaultsQuery = () => {
