+++ core/ui/vault/mutations/useUpdateVaultMutation.ts
@@ -2,18 +2,16 @@ import { vaultsQueryKey } from '@core/ui/query/keys'
 import { useInvalidateQueries } from '@lib/ui/query/hooks/useInvalidateQueries'
 import { useMutation, UseMutationOptions } from '@tanstack/react-query'
 
-import {
-  UpdateVaultFunction,
-  UpdateVaultInput,
-  useUpdateVault,
-} from '../state/updateVault'
+import { UpdateVaultFunction } from '../../state/storage/write'
+import { UpdateVaultInput } from '../../state/storage/write'
+import { useCoreWriteStorage } from '../../state/storage/write'
 
 export const useUpdateVaultMutation = (
   options?: UseMutationOptions<any, any, UpdateVaultInput, unknown>
 ) => {
   const invalidateQueries = useInvalidateQueries()
 
-  const updateVault = useUpdateVault()
+  const { updateVault } = useCoreWriteStorage()
 
   const mutationFn: UpdateVaultFunction = async input => {
     const result = await updateVault(input)
